Add files via upload
This is my monthly expense/ budget tracker. The excel sheet consists of 2 worksheets: data (consists of the expense data table) and dashboard (a visual representation of the analysis of my data). The dashboard clearly displays the total amount of money spent in different categories during the time duration entered. It encompasses of a category amount table, different icons, a bar chart and a pie chart.